// Routing/Buyers/Types.tsx
export interface SellersData {
    Bank: any[]; // Assuming Bank is an array of any type
    LastName: string;
    Product: any[]; // Assuming Product is an array of any type
    address: string;
    businessDescription: string;
    businessName: string;
    businessType: string;
    country: string;
    createdAt: string;
    email: string;
    firstName: string;
    googleID: string;
    hearAboutUs: string;
    id: string;
    isEmailVerified: boolean;
    isEnabled: boolean;
    isPhoneVerified: boolean;
    phoneNumber: string;
    point: string;
    profileImage: any; // Assuming profileImage is of any type
    referralCode: string;
    resetPasswordExpires: any; // Assuming resetPasswordExpires is of any type
    resetPasswordToken: any; // Assuming resetPasswordToken is of any type
    role: string;
    sellerStatus: string;
    subscription: any[]; // Assuming subscription is an array of any type
    supplyCapacity: string;
    totalAnnualRevenue: string;
    updatedAt: string;
    yearEstablished: string;
    message: string;
    type: string;
    messageID?: string
 
  }
  